Move lobbyJoinErrorMessage to a leaf module to preserve the lobby-view code-split
join-lobby.tsx loads eagerly via PlayRoot while the lobby view is
React.lazy (and web builds never route to it), so importing the helper from
./view pulled the whole in-lobby UI into the main chunk.
