+++ client/lobbies/view.tsx
@@ -1,4 +1,3 @@
-import { TFunction } from 'i18next'
 import { InvokeError } from 'nydus-client'
 import * as React from 'react'
 import { useEffect, useState } from 'react'
@@ -41,6 +40,7 @@ import {
   startCountdown,
 } from './action-creators'
 import LobbyComponent from './lobby'
+import { lobbyJoinErrorMessage } from './lobby-join-errors'
 import { LobbySummaryDetails, LobbySummaryLoadState, useLobbySummary } from './lobby-summary'
 import { useCorrectLobbySlug } from './lobby-url'
 
@@ -355,34 +355,6 @@ const JoinPreviewLayout = styled.div`
   flex-direction: column;
   align-items: center;
 `
-
-/**
- * Returns the user-facing message for a failed lobby join. Shared by every UI that dispatches
- * `joinLobby` so the same failure reads the same everywhere.
- */
-export function lobbyJoinErrorMessage(err: unknown, t: TFunction): string {
-  const code = err instanceof InvokeError ? err.body?.code : undefined
-  switch (code) {
-    case LobbyJoinErrorCode.NoLongerOpen:
-      return t('lobbies.summary.noLongerOpen', 'This lobby is no longer open.')
-    case LobbyJoinErrorCode.Full:
-      return t('lobbies.joinLobby.errorFull', 'This lobby is full.')
-    case LobbyJoinErrorCode.Banned:
-      return t('lobbies.joinLobby.errorBanned', "You've been banned from this lobby.")
-    case LobbyJoinErrorCode.AlreadyStarted:
-      return t('lobbies.state.started', 'This lobby has already started and cannot be joined.')
-    case LobbyJoinErrorCode.AlreadyInActivity:
-      return t(
-        'lobbies.joinLobby.errorAlreadyInActivity',
-        'You are already in a game, searching for a match, or in another lobby.',
-      )
-    default:
-      return t(
-        'lobbies.joinLobby.errorGeneric',
-        'Something went wrong while joining the lobby. Please try again.',
-      )
-  }
-}
 
 /**
  * Renders the join interstitial for a lobby that isn't the current user's active lobby: a preview
